move AccessInfo class to separate file
This fixes a rather strange compile error: If AccessInfo is declared in
FindStaticFieldAccesses.java, compilation fails with the reason
'symbol not found: class CGConsumer'. This happens both with a standard
oracle java 1.7 and 1.8 compiler but (very strangely) not in eclipse.
This error can be fixed by making AccessInfo an inner class or by moving
it into a separate file. Why? I don't know :-(